Android example app package rename
