+++ example/android/app/build.gradle
@@ -128,7 +128,7 @@ android {
     }
 
     defaultConfig {
-        applicationId "com.example"
+        applicationId "com.adjust.examples"
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
         versionCode 1
