+++ example/ios/Podfile
@@ -34,6 +34,7 @@ target 'example' do
   pod 'glog', :podspec => '../node_modules/react-native/third-party-podspecs/glog.podspec'
   pod 'Folly', :podspec => '../node_modules/react-native/third-party-podspecs/Folly.podspec'
 
+
   pod 'react-native-adjust', :path => '../node_modules/react-native-adjust'
 
   target 'exampleTests' do
